requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
celery==5.3.4
dj-database-url
drf-spectacular 
requests
beautifulsoup4
django-redis == 5.2
django-celery-beat>=2.5.0
django-celery-results>=2.4.0
